refactor: 러닝, 비러닝 데이터 model 디렉토리 이동
