+++ server/service.py
@@ -1,0 +1,10 @@
+from model.running_model import *
+
+
+def main():
+    model = Model()
+    running_data = model.create_model()
+
+
+if __name__ == "__main__":
+    main()
